requirements.txt
langchain>=0.0.216
numpy>=1.25.0
openai<=0.27.8
pygame-gui>=0.6.9
pygame>=2.4.0
python-dotenv>=1.0.0
